[tf.data] Cosmetic changes to tf.data optimizers: - Use graph_utils::GetInputNode in optimizers - Moved python optimization tests into their own files from optimize_dataset_op_test
PiperOrigin-RevId: 209819734
